Improve diff handling
previously when there was a bucket with multiple matches, it would keep them all in a single result instance.
this led to problems when trying to use checkRemove to filter out differences which would affect some but not all rows in a given result item.

Changed so that if the number or rows in a bucket are the same between from and to, to split those up into separate result instances so they can be removed separately.